Update import process and enhance product touch feature
Refactored the import process in class-boekdb-import.php to improve its readability and maintainability. Added more documentation to clarify the purpose and functionality of several methods and class variables. Enhanced the touch product feature in the class-boekdb-post-types.php, which now includes a new touch product button on the posts dashboard for "boekdb_boek" post type. Lastly, Yoda condition checking is now disabled in the phpcs.xml file to align with code style preferences.